Bit of refactoring/cleanup
