+++ lib/store/types.ts
@@ -18,10 +18,10 @@ import {
 import { BrandProperties } from "@lib/types";
 
 import { type Indexes } from "@lib/utils/form-builder/getPath";
-import { transformFormPropertiesOptions } from "@lib/store/helpers/elements/transformFormProperties";
 
 export interface TemplateStoreState extends TemplateStoreProps {
   focusInput: boolean;
+  setHasTransformed: () => void;
   setHasHydrated: () => void;
   getFocusInput: () => boolean;
   moveUp: (index: number, groupId?: string) => void;
@@ -72,8 +72,8 @@ export interface TemplateStoreState extends TemplateStoreProps {
   unsetField: (path: string) => void;
   duplicateElement: (id: number, groupId?: string, copyEn?: string, copyFr?: string) => void;
   importTemplate: (jsonConfig: FormProperties) => void;
-  transform: (options?: transformFormPropertiesOptions) => void;
-  getSchema: (options?: transformFormPropertiesOptions) => string;
+  transform: () => void;
+  getSchema: () => string;
   getIsPublished: () => boolean;
   setIsPublished: (isPublished: boolean) => void;
   getFormElementById: (id: number) => FormElement | undefined;
@@ -105,6 +105,7 @@ export interface TemplateStoreProps {
   translationLanguagePriority: Language;
   focusInput: boolean;
   hasHydrated: boolean;
+  hasTransformed: boolean;
   form: FormProperties;
   isPublished: boolean;
   name: string;
