+++ lib/store/useTemplateStore.tsx
@@ -53,9 +53,7 @@ import {
   localizeField,
   getFormElementIndexes,
 } from "./helpers/elements";
-import { transformFormProperties } from "@lib/store/helpers/elements/transformFormProperties";
-
-import { transform } from "./helpers/transform/transform";
+import { transform } from "./helpers/elements/transformFormProperties";
 
 const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => {
   const props = initStore(initProps);
@@ -113,15 +111,16 @@ const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => 
             getHighestElementId: getHighestElementId(set, get),
             generateElementId: generateElementId(set, get),
             transform: transform(set, get),
-            getSchema: (options) => {
-              const schema = getSchemaFromState(get(), get().allowGroupsFlag);
-              const transformedSchema = transformFormProperties(schema, options);
-
+            getSchema: () => {
+              // hasHydrated should work here but we get an error. leaving this timeout for now.
               setTimeout(() => {
-                get().transform();
+                if (!get().hasTransformed) {
+                  get().transform();
+                }
+                set({ hasTransformed: true });
               }, 100);
 
-              return JSON.stringify(transformedSchema, null, 2);
+              return JSON.stringify(getSchemaFromState(get(), get().allowGroupsFlag), null, 2);
             },
             getId: () => get().id,
             getIsPublished: () => get().isPublished,
@@ -149,6 +148,7 @@ const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => 
             },
             updateSecurityAttribute: (value) => set({ securityAttribute: value }),
             resetDeliveryOption: () => set({ deliveryOption: undefined }),
+            setHasTransformed: () => set({ hasTransformed: true }),
           }),
 
           storageOptions
